+++ src/Handler/AeatClientHandler.php
@@ -78,6 +78,15 @@ final readonly class AeatClientHandler
         $client->setCertificate($this->aeatClientConfig['pfx_certificate_filepath'], $this->aeatClientConfig['pfx_certificate_password']);
         $client->setEntitySeal($this->aeatClientConfig['is_entity_seal_certificate']);
         $client->setProduction($this->aeatClientConfig['is_prod_environment']);
+        if (null !== ($this->aeatClientConfig['representative'] ?? null)) {
+            $client->setRepresentative($this->fiscalIdentifierFactory->makeValidatedFiscalIdentifierModelFromConfigArray($this->aeatClientConfig['representative']));
+        }
+        if (null !== $this->aeatClientConfig['requirement_reference']) {
+            $client->setRequirementReference($this->aeatClientConfig['requirement_reference'], $this->aeatClientConfig['requirement_is_last_submission']);
+        }
+        if (null !== $this->aeatClientConfig['voluntary_remission_end_date']) {
+            $client->setVoluntaryRemissionEndDate(new \DateTimeImmutable($this->aeatClientConfig['voluntary_remission_end_date']), $this->aeatClientConfig['voluntary_remission_is_affected_by_incident']);
+        }
 
         return $client;
     }